Refactor Rust plugin
