--- lib/resolver/rust-crate.js
@@ -1,5 +1,0 @@
-import liveResolverQuery from './live-resolver-query.js';
-
-export default function ({ target }) {
-  return liveResolverQuery({ type: 'crates', target });
-}
