+++ test/plugins/rust.test.js
@@ -1,10 +1,10 @@
 import assert from 'assert';
-import rustCrate from '../../lib/resolver/rust-crate.js';
+import rustCrate from '../../lib/plugins/rust';
 
 describe('rust-crate', () => {
   it('resolves hamcrest using the live-resolver', () => {
     assert.deepEqual(
-      rustCrate({ target: 'hamcrest' }),
+      rustCrate.resolve({ target: 'hamcrest' }),
       {
         url: 'https://githublinker.herokuapp.com/q/crates/hamcrest',
         method: 'GET',
